server/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==1.10.9
